codebaseB/smart_cane/buttons/requirements.txt
paho-mqtt
RPi.GPIO 